add aadjust / refactor
